refactor(builder): share Dragonborn lineage table with ChoicePicker (#292)
Review follow-up: the sheet's PendingChoicesPanel renders lineage choices via
ChoicePicker, which still showed a flat radio list of bare colors — inconsistent
with the new builder table and the header's kind qualifier. Extracted the table
into a shared DragonbornLineageTable component and use it in both LineagePicker
and ChoicePicker's lineage branch for dragonborn. Adds ChoicePicker tests.
